refactor(desktop): 归并既有 settings panel 至 settings/panels
将 agents、hooks、integrations 设置面板迁入 settings/panels/，agents 面板改用共享 SettingsRow。
