+++ apps/desktop/src/components/settings/panels/hooks-settings-panel.tsx
