+++ apps/desktop/src/components/settings/panels/integrations-settings-panel.tsx
